# 17_USA/openapi.json
{"openapi":"3.0.2","info":{"title":"Naval Ship API","version":"1.0"},"paths":{"/api/v1/user/{user_id}":{"get":{"tags":["user"],"summary":"Fetch User","description":"Fetch a user by ID","operationId":"fetch_user_api_v1_user__user_id__get","parameters":[{"required":true,"schema":{"title":"User Id","type":"integer"},"name":"user_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/user/signup":{"post":{"tags":["user"],"summary":"Create User Account","description":"Create new user","operationId":"create_user_signup_api_v1_user_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignup"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/user/login":{"post":{"tags":["user"],"summary":"User Login","description":"Logs in the user and returns JWT","operationId":"login_api_v1_user_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_api_v1_user_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/user/updatepassword":{"post":{"tags":["user"],"summary":"Update User Password","description":"Update a user password","operationId":"update_password_api_v1_user_updatepassword_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPWDchange"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/admin/":{"get":{"tags":["admin"],"summary":"Admin Check","description":"Returns true if the user is in admin role","operationId":"admin_check_api_v1_admin__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/admin/file":{"post":{"tags":["admin"],"summary":"Get File","description":"Returns a file on the server","operationId":"get_file_api_v1_admin_file_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetFile"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/admin/getFlag":{"put":{"tags":["admin"],"summary":"Get Flag","description":"The Flag","operationId":"get_flag_api_v1_admin_getFlag_put","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"Body_login_api_v1_user_login_post":{"title":"Body_login_api_v1_user_login_post","required":["username","password"],"type":"object","properties":{"grant_type":{"title":"Grant Type","pattern":"password","type":"string"},"username":{"title":"Username","type":"string"},"password":{"title":"Password","type":"string"},"scope":{"title":"Scope","type":"string","default":""},"client_id":{"title":"Client Id","type":"string"},"client_secret":{"title":"Client Secret","type":"string"}}},"GetFile":{"title":"GetFile","required":["file"],"type":"object","properties":{"file":{"title":"File","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Token":{"title":"Token","required":["access_token","token_type"],"type":"object","properties":{"access_token":{"title":"Access Token","type":"string"},"token_type":{"title":"Token Type","type":"string"}}},"User":{"title":"User","type":"object","properties":{"guid":{"title":"Guid","type":"string"},"email":{"title":"Email","type":"string","format":"email"},"date":{"title":"Date","type":"integer"},"time_created":{"title":"Time Created","type":"integer"},"admin":{"title":"Admin","type":"boolean","default":false},"id":{"title":"Id","type":"integer"}},"description":"Utilized for authentication."},"UserPWDchange":{"title":"UserPWDchange","required":["guid","password"],"type":"object","properties":{"guid":{"title":"Guid","type":"string"},"password":{"title":"Password","type":"string"}}},"UserSignup":{"title":"UserSignup","required":["email","password"],"type":"object","properties":{"email":{"title":"Email","type":"string","format":"email"},"password":{"title":"Password","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/user/login"}}}}}}